Stop polling moodle after user has been logged out
Without this change, if a user logs out in another tab, or their sesskey is
otherwise made invalid, then RTQ will keep polling Moodle indefinitely.

Once we see an invalidsesskey or requirelogin error, the requests from RTQ are
going to work until the page is reloaded (so the requests use the new sesskey
if the user has since logged back in).